+++ src/StateSmith/Output/Sim/SimWebGenerator.cs
@@ -1,8 +1,11 @@
+#nullable enable
+
 using StateSmith.Input.Expansions;
 using StateSmith.Output.UserConfig;
 using StateSmith.Runner;
 using StateSmith.SmGraph;
 using System.IO;
+using System.Text;
 
 namespace StateSmith.Output.Sim;
 
@@ -10,8 +13,7 @@ public class SimWebGenerator
 {
     public RunnerSettings RunnerSettings => runner.Settings;
 
-    string diagramPath;
-    private readonly string outputDir;
+    private readonly ICodeFileWriter codeFileWriter;
     MermaidEdgeTracker mermaidEdgeTracker = new();
     TrackingExpander trackingExpander = new();
     TextWriter mermaidCodeWriter = new StringWriter();
@@ -20,38 +22,44 @@ public class SimWebGenerator
     StateMachineProvider stateMachineProvider;
     SmRunner runner;
 
-    public SimWebGenerator(string diagramPath, string outputDir)
+    public SimWebGenerator(ICodeFileWriter codeFileWriter)
     {
-        this.diagramPath = diagramPath;
-        this.outputDir = outputDir;
-        DiServiceProvider diServiceProvider;
+        // Internally, the `SimWebGenerator` uses a `SmRunner` to transform the input diagram into a simulation web page.
+        // To customize the transformation/code generation process, we register custom DI services with the `SmRunner`.
 
-        runner = new(diagramPath: diagramPath, renderConfig: new SimRenderConfig(), transpilerId: TranspilerId.JavaScript);
+        this.codeFileWriter = codeFileWriter;
+        DiServiceProvider simDiServiceProvider;
+
+        runner = new(diagramPath: "placeholder-updated-in-generate-method.txt", renderConfig: new SimRenderConfig(), transpilerId: TranspilerId.JavaScript);
         runner.Settings.propagateExceptions = true;
 
         // Registering DI services must be done before accessing `runner.SmTransformer`.
-        diServiceProvider = runner.GetExperimentalAccess().DiServiceProvider;
-        diServiceProvider.AddSingletonT<IExpander>(trackingExpander);
-        diServiceProvider.AddSingletonT<ICodeFileWriter>(fileCapturer);
-        diServiceProvider.AddSingletonT<IConsolePrinter>(new DiscardingConsolePrinter());
+        simDiServiceProvider = runner.GetExperimentalAccess().DiServiceProvider;
+        simDiServiceProvider.AddSingletonT<IExpander>(trackingExpander);
+        simDiServiceProvider.AddSingletonT<ICodeFileWriter>(fileCapturer);
+        simDiServiceProvider.AddSingletonT<IConsolePrinter>(new DiscardingConsolePrinter());   // we want regular SmRunner console output to be discarded
 
         // Note! For `MermaidEdgeTracker` to function correctly, both below transformations must occur in the same `SmRunner`.
         // This allows us to easily map an SS behavior to its corresponding mermaid edge ID.
         runner.SmTransformer.InsertBeforeFirstMatch(StandardSmTransformer.TransformationId.Standard_RemoveNotesVertices, new TransformationStep(id: "some string id", GenerateMermaidCode));
         runner.SmTransformer.InsertBeforeFirstMatch(StandardSmTransformer.TransformationId.Standard_Validation1, new TransformationStep(id: "my custom step blah", V1LoggingTransformationStep));
-        stateMachineProvider = diServiceProvider.GetInstanceOf<StateMachineProvider>();
+        stateMachineProvider = simDiServiceProvider.GetInstanceOf<StateMachineProvider>();
     }
 
-    public void Generate()
+    public void Generate(string diagramPath, string outputDir)
     {
+        runner.Settings.DiagramPath = diagramPath;
         runner.Run();
         var smName = stateMachineProvider.GetStateMachine().Name;
 
         if (Directory.Exists(outputDir) == false)
             Directory.CreateDirectory(outputDir);
 
-        using StreamWriter htmlWriter = new(Path.Combine(outputDir, $"{smName}.sim.html"));
-        HtmlRenderer.Render(htmlWriter, smName: smName, mocksCode: mocksWriter.ToString(), mermaidCode: mermaidCodeWriter.ToString(), jsCode: fileCapturer.CapturedCode);
+        string path = Path.Combine(outputDir, $"{smName}.sim.html");
+
+        var sb = new StringBuilder();
+        HtmlRenderer.Render(sb, smName: smName, mocksCode: mocksWriter.ToString(), mermaidCode: mermaidCodeWriter.ToString(), jsCode: fileCapturer.CapturedCode);
+        codeFileWriter.WriteFile(path, code: sb.ToString());
     }
 
     void GenerateMermaidCode(StateMachine sm)