+++ src/StateSmith/Runner/SimulationSettings.cs
@@ -4,15 +4,12 @@ namespace StateSmith.Runner;
 
 public class SimulationSettings
 {
-    public bool enableGeneration = true;
+    public bool enableGeneration = false;
 
     /// <summary>
     /// Defaults to same directory as <see cref="RunnerSettings.outputDirectory"/>.
     /// </summary>
     public string? outputDirectory = null;
-
-    /// <summary>
-    /// Defaults to state machine name + ".sim.html"
-    /// </summary>
-    public string? outputFileName = null;
+    
+    public string outputFileNamePostfix = ".sim.html";
 }